pyproject.toml
[project]
name = "imprint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "dependency-injector>=4.48.3",
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "simhash>=2.1.2",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
